requirements.txt
fastapi 
python-multipart 
psycopg2-binary 
python-jose 
passlib[bcrypt]
uvicorn 
sqlalchemy